feat(cli): added new command generate, to help generate files in the project chore(): updated dependencies refactor(core): changed order or HttpClientTypeSchema feat(core): added warning message when trying to validate or map a body on a GET or DELETE end-point refactor(): replaced all z.union of literal with z.enum feat(cli): added check to verify if the command is running on a API BFF project chore(cli): removed ts-expect-error comment from tsup.config.ts

+++ packages/cli/src/dev.ts
@@ -8,6 +8,7 @@ import { rimraf } from 'rimraf';
 import { typechecking_plugin } from '../typechecking-plugin.js';
 
 import { get_default_esbuild_options } from './get-default-esbuild-options.js';
+import { is_api_bff_project } from './is-api-bff-project.js';
 
 function dev_plugin(): Plugin {
   return {
@@ -63,6 +64,11 @@ function dev_plugin(): Plugin {
 export const dev_command = new Command('dev')
   .description('Development server')
   .action(async () => {
+    const isApiBffProject = await is_api_bff_project(process.cwd());
+    if (!isApiBffProject) {
+      console.error('Not in a API BFF Project');
+      return;
+    }
     const [defaultOptions] = await Promise.all([
       get_default_esbuild_options(),
       rimraf('dist'),
